eval/OpenRT/example/test_openai_image.py
import sys
import os
from pathlib import Path
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Add the parent directory to the path so we can import the module
sys.path.append(str(Path(__file__).parent.parent))

try:
    from OpenRT.models.implementations.openai_model import OpenAIModel
    from PIL import Image
    
    # Create a test image
    img = Image.new('RGB', (100, 100), color='red')
    img.save("test_image.png")
    
    # Initialize the model (replace with your API key)
    model = OpenAIModel(
        api_key=os.getenv("OPENAI_API_KEY"),
        base_url=os.getenv("OPENAI_BASE_URL"),
        model_name="gpt-4-vision-preview"
    )
    
    # Test image query
    response = model.query(
        text_input="What color is this square? Answer with a single word.",
        image_input="test_image.png"
    )
    
    print(f"Response: {response}")
    os.remove("test_image.png")
    
except ImportError as e:
    print(f"Import error: {e}")
except Exception as e:
    print(f"Error: {e}")
    if os.path.exists("test_image.png"):
        os.remove("test_image.png")
